Fixed my array and values to be doubles rather than ints so that it displays correctly. My shoulder just slowed me down to much in class and I did not have class time to finish the change but wanted to see it complete

+++ CSC390Midterm WatchKit Extension/InterfaceController.swift
@@ -14,13 +14,13 @@ class InterfaceController: WKInterfaceController {
 
     @IBOutlet var listpicker: WKInterfacePicker!
     
-    @IBAction func controlbutton1(value : Int)
+    @IBAction func controlbutton1(value : Double)
     {
         conversioncore.speedselected = value
         print("\(conversioncore.speedselected)")
         
     }
-    @IBAction func pickerupdate(value: Int)
+    @IBAction func pickerupdate(value: Double)
     {
         conversioncore.speedselected = value
         print("\(conversioncore.speedselected)")
